CScene_Viewer 제작
파일열기후 파일선택 그다음 상단 메뉴탭에서 애니메이션 에서 설정 적용시 애니메이션 실행

+++ Client/Gun.cpp
@@ -90,19 +90,19 @@ void Gun::Shot()
 
 	Vect2 vShotPos = GetPos();
 
-	for (UINT i = 0; i < mInfo.shotCount; ++i)
+	for (int i = 0; i < mInfo.shotCount; ++i)
 	{
 		// 일정 발사각 범위 내의 랜덤한 방향을 생성합니다.
-		int shotAngle = (int)(mInfo.shotAngle * 0.5f);
+		int shotAngle = mInfo.shotAngle * 0.5f;
 		float angle = GetAngleOrg() + (float)CRandom::GetI()->Next(shotAngle * -1, shotAngle); // 랜덤한 각도
 		Vect2 vDir = Vect2::FromAngle(angle);
 
 		// 총알 오브젝트
 		tBullet bInfo = {};
 		bInfo.damage = mInfo.atkDamage;
-		bInfo.bounceCount = (float)mInfo.bouncCount;
-		bInfo.penetrationCount = (float)mInfo.penetration;
-		bInfo.splitCount = (float)mInfo.splitCount;
+		bInfo.bounceCount = mInfo.bouncCount;
+		bInfo.penetrationCount = mInfo.penetration;
+		bInfo.splitCount = mInfo.splitCount;
 
 		Bullet* pBullet = new Bullet(mInfo.bulletType);
 		pBullet->SetPos(vShotPos + vDir.Normalize() * 35.f);
